+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Improved reporting of local ID parsing errors.
+- Replace `fetch-object-as` and `fetch-portable-object` commands with `fetch-object` command.
 
 ### Fixed